fix:    Separate extension description into separate files.
