+++ Source/WebPImage.swift
@@ -32,26 +32,3 @@ internal func _decode(_ data: Data) -> Image? {
         return WebPDataDecoder.decode(data)
     }
 }
-
-// MARK: - WebP Format Testing
-extension Data {
-    // Borrow from KingfisherWebp
-    // https://github.com/Yeatse/KingfisherWebP/blob/master/KingfisherWebP/Classes/Image%2BWebP.swift#L38
-    var isWebPFormat: Bool {
-        if count < 12 {
-            return false
-        }
-        
-        let endIndex = index(startIndex, offsetBy: 12)
-        let testData = subdata(in: startIndex..<endIndex)
-        guard let testString = String(data: testData, encoding: .ascii) else {
-            return false
-        }
-        
-        if testString.hasPrefix("RIFF") && testString.hasSuffix("WEBP") {
-            return true
-        } else {
-            return false
-        }
-    }
-}
